requirements.txt
flask
flask-cors
pillow
requests
python-dotenv
gunicorn
